REFACTOR: ContentRepository -> ContentQueryRepository
