+++ api/baeuja/src/main/java/xyz/baeuja/api/content/repository/ContentRepository.java
@@ -4,9 +4,7 @@ import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import xyz.baeuja.api.content.domain.Content;
-import xyz.baeuja.api.home.dto.HomeContentResponse;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
@@ -34,21 +32,5 @@ public class ContentRepository {
      */
     public Optional<Content> findOne(long id) {
         return Optional.ofNullable(em.find(Content.class, id));
-    }
-
-    /**
-     * home 에서 가장 최근에 추가된 content 를 조회. JPQL 로 SentenceWord 까지 조인
-     *
-     * @return content 정보와 unit 개수, 단어 개수를 포함한 HomeContentResponse 리스트
-     */
-    public List<HomeContentResponse> findTop10ByOrderByCreatedAtDesc() {
-        return em.createQuery("select new xyz.baeuja.api.home.dto.HomeContentResponse(" +
-                        "c.classification, c.title, c.artist, c.director, c.thumbnailUrl, " +
-                        "(select count(u) from Unit u where u.content.id = c.id), " +
-                        "(select count(sw.word) from SentenceWord sw where sw.sentence.unit.content.id = c.id))" +
-                        "from Content c " +
-                        "order by c.createdAt desc", HomeContentResponse.class)
-                .setMaxResults(10)
-                .getResultList();
     }
 }
