+++ api/baeuja/src/test/java/xyz/baeuja/api/content/repository/query/ContentQueryRepositoryTest.java
@@ -1,6 +1,5 @@
-package xyz.baeuja.api.content.repository;
+package xyz.baeuja.api.content.repository.query;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,7 +22,19 @@ import static org.junit.jupiter.api.Assertions.*;
         "/sql/sentence_word.sql"},
         executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
 @SpringBootTest
-class ContentRepositoryTest {
+class ContentQueryRepositoryTest {
 
-    @Autowired ContentRepository contentRepository;
+    @Autowired ContentQueryRepository contentRepository;
+
+    @Test
+    @DisplayName("최근에 추가된 10개 content 정보 조회")
+    void findTop10ByOrderByCreatedAtDesc_success() {
+        // when
+        List<HomeContentResponse> latestContents = contentRepository.findTop10ByOrderByCreatedAtDesc();
+
+        // then
+        assertThat(latestContents).hasSize(10);
+        assertThat(latestContents.get(0).getUnitCount()).isNotZero();
+        assertThat(latestContents.get(0).getWordCount()).isNotZero();
+    }
 }
